+++ spec/unit/coinbase/balance_map_spec.rb
@@ -1,34 +1,46 @@
 # frozen_string_literal: true
 
 describe Coinbase::BalanceMap do
+  let(:eth_asset) do
+    Coinbase::Client::Asset.new(network_id: 'base-sepolia', asset_id: 'eth', decimals: 18)
+  end
+
   describe '.from_balances' do
     let(:eth_amount) { BigDecimal('123.0') }
-    let(:eth_asset) { instance_double('Coinbase::Client::Asset', asset_id: 'ETH') }
     let(:eth_balance_model) { instance_double('Coinbase::Client::Balance', asset: eth_asset, amount: eth_amount) }
+    let(:eth_atomic_amount) { Coinbase::Asset.from_model(eth_asset).from_atomic_amount(eth_amount) }
 
     let(:usdc_amount) { BigDecimal('456.0') }
-    let(:usdc_asset) { instance_double('Coinbase::Client::Asset', asset_id: 'USDC') }
+    let(:usdc_asset) do
+      Coinbase::Client::Asset.new(network_id: 'base-sepolia', asset_id: 'usdc', decimals: 6)
+    end
     let(:usdc_balance_model) { instance_double('Coinbase::Client::Balance', asset: usdc_asset, amount: usdc_amount) }
+    let(:usdc_atomic_amount) { Coinbase::Asset.from_model(usdc_asset).from_atomic_amount(usdc_amount) }
 
     let(:weth_amount) { BigDecimal('789.0') }
     let(:weth_asset) { instance_double('Coinbase::Client::Asset', asset_id: 'WETH') }
+    let(:weth_asset) do
+      Coinbase::Client::Asset.new(network_id: 'base-sepolia', asset_id: 'weth', decimals: 18)
+    end
     let(:weth_balance_model) { instance_double('Coinbase::Client::Balance', asset: weth_asset, amount: weth_amount) }
+    let(:weth_atomic_amount) { Coinbase::Asset.from_model(weth_asset).from_atomic_amount(weth_amount) }
 
     let(:balances) { [eth_balance_model, usdc_balance_model, weth_balance_model] }
 
     subject { described_class.from_balances(balances) }
 
     it 'returns a new BalanceMap object with the correct balances' do
-      expect(subject[:eth]).to eq(eth_amount / BigDecimal(Coinbase::WEI_PER_ETHER))
-      expect(subject[:usdc]).to eq(usdc_amount / BigDecimal(Coinbase::ATOMIC_UNITS_PER_USDC))
-      expect(subject[:weth]).to eq(weth_amount / BigDecimal(Coinbase::WEI_PER_ETHER))
+      expect(subject[:eth]).to eq(eth_atomic_amount)
+      expect(subject[:usdc]).to eq(usdc_atomic_amount)
+      expect(subject[:weth]).to eq(weth_atomic_amount)
     end
   end
 
   describe '#add' do
     let(:amount) { BigDecimal('123.0') }
     let(:asset_id) { :eth }
-    let(:balance) { Coinbase::Balance.new(amount: amount, asset_id: asset_id) }
+    let(:asset) { Coinbase::Asset.from_model(eth_asset) }
+    let(:balance) { Coinbase::Balance.new(amount: amount, asset: asset) }
 
     subject { described_class.new }
 
@@ -39,7 +51,7 @@ describe Coinbase::BalanceMap do
     end
 
     context 'when the balance is not a Coinbase::Balance' do
-      let(:balance) { instance_double('Coinbase::Balance') }
+      let(:balance) { instance_double('Coinbase::Asset') }
 
       it 'raises an ArgumentError' do
         expect { subject.add(balance) }.to raise_error(ArgumentError)
@@ -50,7 +62,8 @@ describe Coinbase::BalanceMap do
   describe '#to_s' do
     let(:amount) { BigDecimal('123.0') }
     let(:asset_id) { :eth }
-    let(:balance) { Coinbase::Balance.new(amount: amount, asset_id: asset_id) }
+    let(:asset) { Coinbase::Asset.from_model(eth_asset) }
+    let(:balance) { Coinbase::Balance.new(amount: amount, asset: asset) }
 
     let(:expected_result) { { eth: '123' }.to_s }