+++ spec/unit/coinbase/balance_spec.rb
@@ -1,75 +1,66 @@
 # frozen_string_literal: true
 
 describe Coinbase::Balance do
-  describe '.from_model' do
-    let(:amount) { BigDecimal('123.0') }
-    let(:balance_model) { instance_double('Coinbase::Client::Balance', asset: asset, amount: amount) }
+  let(:amount) { BigDecimal('123.0') }
+  let(:balance_model) { instance_double('Coinbase::Client::Balance', asset: asset, amount: amount) }
+  let(:eth_asset) do
+    Coinbase::Client::Asset.new(network_id: 'base-sepolia', asset_id: 'eth', decimals: 18)
+  end
 
+  describe '.from_model' do
     subject(:balance) { described_class.from_model(balance_model) }
 
     context 'when the asset is :eth' do
-      let(:asset) { instance_double('Coinbase::Client::Asset', asset_id: 'ETH') }
+      let(:asset) { eth_asset }
 
-      it 'returns a new Balance object with the correct amount' do
-        expect(balance.amount).to eq(amount / BigDecimal(Coinbase::WEI_PER_ETHER))
+      it 'returns a Balance object' do
+        expect(balance).to be_a(described_class)
       end
 
-      it 'returns a new Balance object with the correct asset_id' do
+      it 'sets the correct amount' do
+        expect(balance.amount).to eq(amount / BigDecimal(10).power(eth_asset.decimals))
+      end
+
+      it 'sets the correct asset_id' do
         expect(balance.asset_id).to eq(:eth)
       end
     end
 
-    context 'when the asset is :usdc' do
-      let(:asset) { instance_double('Coinbase::Client::Asset', asset_id: 'USDC') }
-
-      it 'returns a new Balance object with the correct amount' do
-        expect(balance.amount).to eq(amount / BigDecimal(Coinbase::ATOMIC_UNITS_PER_USDC))
+    context 'when the asset is other' do
+      let(:decimals) { 9 }
+      let(:asset) do
+        Coinbase::Client::Asset.new(
+          network_id: 'base-sepolia',
+          asset_id: 'other',
+          decimals: decimals
+        )
       end
 
-      it 'returns a new Balance object with the correct asset_id' do
-        expect(balance.asset_id).to eq(:usdc)
-      end
-    end
-
-    context 'when the asset is :weth' do
-      let(:asset) { instance_double('Coinbase::Client::Asset', asset_id: 'WETH') }
-
-      it 'returns a new Balance object with the correct amount' do
-        expect(balance.amount).to eq(amount / BigDecimal(Coinbase::WEI_PER_ETHER))
+      it 'returns a Balance object' do
+        expect(balance).to be_a(described_class)
       end
 
-      it 'returns a new Balance object with the correct asset_id' do
-        expect(balance.asset_id).to eq(:weth)
-      end
-    end
-
-    context 'when the asset is another asset type' do
-      let(:asset) { instance_double('Coinbase::Client::Asset', asset_id: 'OTHER') }
-
-      it 'returns a new Balance object with the correct amount' do
-        expect(balance.amount).to eq(amount)
+      it 'sets the correct amount' do
+        expect(balance.amount).to eq(amount / BigDecimal(10).power(decimals))
       end
 
-      it 'returns a new Balance object with the correct asset_id' do
+      it 'sets the correct asset_id' do
         expect(balance.asset_id).to eq(:other)
       end
     end
   end
 
   describe '.from_model_and_asset_id' do
-    let(:amount) { BigDecimal('123.0') }
-    let(:balance_model) { instance_double('Coinbase::Client::Balance', asset: asset, amount: amount) }
-
     subject(:balance) { described_class.from_model_and_asset_id(balance_model, asset_id) }
 
     context 'when the balance model asset is :eth' do
-      let(:asset) { instance_double('Coinbase::Client::Asset', asset_id: 'ETH') }
+      let(:asset) { eth_asset }
 
       context 'and the specified asset_id is :eth' do
         let(:asset_id) { :eth }
 
         it 'returns a new Balance object with the correct amount' do
-          expect(balance.amount).to eq(amount / BigDecimal(Coinbase::WEI_PER_ETHER))
+          expect(balance.amount).to eq(amount / BigDecimal(10).power(eth_asset.decimals))
         end
 
         it 'returns a new Balance object with the correct asset_id' do
@@ -81,7 +72,7 @@ describe Coinbase::Balance do
         let(:asset_id) { :gwei }
 
         it 'returns a new Balance object with the correct amount' do
-          expect(balance.amount).to eq(amount / BigDecimal(Coinbase::GWEI_PER_ETHER))
+          expect(balance.amount).to eq(amount / BigDecimal(10).power(Coinbase::GWEI_DECIMALS))
         end
 
         it 'returns a new Balance object with the correct asset_id' do
@@ -100,68 +91,69 @@ describe Coinbase::Balance do
           expect(balance.asset_id).to eq(asset_id)
         end
       end
-    end
 
-    context 'when the asset is :usdc' do
-      let(:asset) { instance_double('Coinbase::Client::Asset', asset_id: 'USDC') }
-      let(:asset_id) { :usdc }
+      context 'and the specified asset_id is another asset type' do
+        let(:asset_id) { :other }
 
-      it 'returns a new Balance object with the correct amount' do
-        expect(balance.amount).to eq(amount / BigDecimal(Coinbase::ATOMIC_UNITS_PER_USDC))
-      end
-
-      it 'returns a new Balance object with the correct asset_id' do
-        expect(balance.asset_id).to eq(asset_id)
+        it 'raise an error' do
+          expect { balance }.to raise_error(ArgumentError)
+        end
       end
     end
 
-    context 'when the asset is :weth' do
-      let(:asset) { instance_double('Coinbase::Client::Asset', asset_id: 'WETH') }
-      let(:asset_id) { :weth }
-
-      it 'returns a new Balance object with the correct amount' do
-        expect(balance.amount).to eq(amount / BigDecimal(Coinbase::WEI_PER_ETHER))
-      end
-
-      it 'returns a new Balance object with the correct asset_id' do
-        expect(balance.asset_id).to eq(asset_id)
-      end
-    end
-
-    context 'when the asset is another asset type' do
-      let(:asset) { instance_double('Coinbase::Client::Asset', asset_id: 'OTHER') }
+    context 'when the asset is not eth' do
+      let(:decimals) { 9 }
       let(:asset_id) { :other }
+      let(:asset) do
+        Coinbase::Client::Asset.new(
+          network_id: 'base-sepolia',
+          asset_id: 'other',
+          decimals: decimals
+        )
+      end
 
       it 'returns a new Balance object with the correct amount' do
-        expect(balance.amount).to eq(amount)
+        expect(balance.amount).to eq(amount / BigDecimal(10).power(decimals))
       end
 
       it 'returns a new Balance object with the correct asset_id' do
         expect(balance.asset_id).to eq(asset_id)
+      end
+
+      context 'when the asset ID does not match the asset' do
+        let(:asset_id) { :different }
+
+        it 'raises an error' do
+          expect { balance }.to raise_error(ArgumentError)
+        end
       end
     end
   end
 
   describe '#initialize' do
     let(:amount) { BigDecimal('123.0') }
-    let(:asset_id) { :eth }
+    let(:asset) { Coinbase::Asset.from_model(eth_asset) }
 
-    subject(:balance) { described_class.new(amount: amount, asset_id: asset_id) }
+    subject(:balance) { described_class.new(amount: amount, asset: asset) }
 
     it 'sets the amount' do
       expect(balance.amount).to eq(amount)
     end
 
-    it 'sets the asset_id' do
-      expect(balance.asset_id).to eq(asset_id)
+    it 'sets the asset' do
+      expect(balance.asset).to eq(asset)
+    end
+
+    it "sets the asset_id to the asset's ID" do
+      expect(balance.asset_id).to eq(:eth)
     end
   end
 
   describe '#inspect' do
     let(:amount) { BigDecimal('123.0') }
-    let(:asset_id) { :eth }
+    let(:asset) { Coinbase::Asset.from_model(eth_asset) }
 
-    subject(:balance) { described_class.new(amount: amount, asset_id: asset_id) }
+    subject(:balance) { described_class.new(amount: amount, asset: asset) }
 
     it 'includes balance details' do
       expect(balance.inspect).to include('123', 'eth')